+++ src/configuration/actions_parser.cpp
@@ -15,8 +15,8 @@
 #include "configuration/common/common.hpp"
 #include "configuration/common/configuration.hpp"
 #include "configuration/common/configuration_collector.hpp"
+#include "configuration/common/parser_exception.hpp"
 #include "configuration/common/raw_configuration.hpp"
-#include "exception.hpp"
 #include "log.hpp"
 #include "uri_utils.hpp"
