+++ src/configuration/common/common.hpp
@@ -8,8 +8,8 @@
 
 #include <string>
 
+#include "configuration/common/parser_exception.hpp"
 #include "configuration/common/raw_configuration.hpp"
-#include "exception.hpp"
 #include "ruleset_info.hpp"
 
 using base_section_info = ddwaf::base_ruleset_info::base_section_info;
