+++ src/configuration/common/expression_parser.cpp
@@ -21,9 +21,9 @@
 #include "configuration/common/common.hpp"
 #include "configuration/common/expression_parser.hpp"
 #include "configuration/common/matcher_parser.hpp"
+#include "configuration/common/parser_exception.hpp"
 #include "configuration/common/raw_configuration.hpp"
 #include "configuration/common/transformer_parser.hpp"
-#include "exception.hpp"
 #include "expression.hpp"
 #include "log.hpp"
 #include "matcher/equals.hpp"
