+++ src/configuration/common/raw_configuration.cpp
@@ -13,9 +13,9 @@
 #include <utility>
 #include <vector>
 
+#include "configuration/common/parser_exception.hpp"
 #include "configuration/common/raw_configuration.hpp"
 #include "ddwaf.h"
-#include "exception.hpp"
 #include "semver.hpp"
 #include "utils.hpp"
