+++ src/configuration/common/transformer_parser.cpp
@@ -9,9 +9,9 @@
 #include <vector>
 
 #include "condition/base.hpp"
+#include "configuration/common/parser_exception.hpp"
 #include "configuration/common/raw_configuration.hpp"
 #include "configuration/common/transformer_parser.hpp"
-#include "exception.hpp"
 #include "transformer/base.hpp"
 
 namespace ddwaf {
