+++ src/configuration/legacy_rule_parser.cpp
@@ -19,11 +19,11 @@
 #include "configuration/common/common.hpp"
 #include "configuration/common/configuration.hpp"
 #include "configuration/common/configuration_collector.hpp"
+#include "configuration/common/parser_exception.hpp"
 #include "configuration/common/raw_configuration.hpp"
 #include "configuration/common/transformer_parser.hpp"
 #include "configuration/legacy_rule_parser.hpp"
 #include "ddwaf.h"
-#include "exception.hpp"
 #include "expression.hpp"
 #include "log.hpp"
 #include "matcher/base.hpp"
