+++ src/configuration/processor_parser.cpp
@@ -13,10 +13,10 @@
 #include "configuration/common/configuration.hpp"
 #include "configuration/common/configuration_collector.hpp"
 #include "configuration/common/expression_parser.hpp"
+#include "configuration/common/parser_exception.hpp"
 #include "configuration/common/raw_configuration.hpp"
 #include "configuration/common/reference_parser.hpp"
 #include "configuration/processor_parser.hpp"
-#include "exception.hpp"
 #include "log.hpp"
 #include "processor/base.hpp"
 #include "processor/extract_schema.hpp"
