+++ src/configuration/rule_override_parser.cpp
@@ -12,10 +12,10 @@
 #include "configuration/common/common.hpp"
 #include "configuration/common/configuration.hpp"
 #include "configuration/common/configuration_collector.hpp"
+#include "configuration/common/parser_exception.hpp"
 #include "configuration/common/raw_configuration.hpp"
 #include "configuration/common/reference_parser.hpp"
 #include "configuration/rule_override_parser.hpp"
-#include "exception.hpp"
 #include "log.hpp"
 #include "ruleset_info.hpp"
 #include "uuid.hpp"
