+++ src/configuration/rule_parser.cpp
@@ -14,10 +14,10 @@
 #include "configuration/common/configuration.hpp"
 #include "configuration/common/configuration_collector.hpp"
 #include "configuration/common/expression_parser.hpp"
+#include "configuration/common/parser_exception.hpp"
 #include "configuration/common/raw_configuration.hpp"
 #include "configuration/common/transformer_parser.hpp"
 #include "configuration/rule_parser.hpp"
-#include "exception.hpp"
 #include "log.hpp"
 #include "rule.hpp"
 #include "semver.hpp"
