+++ src/interface.cpp
@@ -99,19 +99,17 @@ ddwaf::waf *ddwaf_init(
             ddwaf::waf_builder builder(
                 limits_from_config(config), free_fn, obfuscator_from_config(config));
 
-            const ddwaf::raw_configuration input = *ruleset;
+            ddwaf::raw_configuration input = *ruleset;
             if (diagnostics == nullptr) {
                 ddwaf::null_ruleset_info ri;
 
-                auto input_map = static_cast<ddwaf::raw_configuration::map>(input);
-                builder.add_or_update("default", input_map, ri);
+                builder.add_or_update("default", input, ri);
                 return new ddwaf::waf{builder.build()};
             }
 
             ddwaf::ruleset_info ri;
             const ddwaf::scope_exit on_exit([&]() { ri.to_object(*diagnostics); });
-            auto input_map = static_cast<ddwaf::raw_configuration::map>(input);
-            builder.add_or_update("default", input_map, ri);
+            builder.add_or_update("default", input, ri);
             return new ddwaf::waf{builder.build()};
         }
     } catch (const std::exception &e) {
@@ -290,16 +288,15 @@ bool ddwaf_builder_add_or_update_config(ddwaf::waf_builder *builder, const char 
 
     try {
         auto input = static_cast<ddwaf::raw_configuration>(*config);
-        auto input_map = static_cast<ddwaf::raw_configuration::map>(input);
 
         if (diagnostics == nullptr) {
             ddwaf::null_ruleset_info ri;
-            return builder->add_or_update({path, path_len}, input_map, ri);
+            return builder->add_or_update({path, path_len}, input, ri);
         }
 
         ddwaf::ruleset_info ri;
         const ddwaf::scope_exit on_exit([&]() { ri.to_object(*diagnostics); });
-        return builder->add_or_update({path, path_len}, input_map, ri);
+        return builder->add_or_update({path, path_len}, input, ri);
     } catch (const std::exception &e) {
         DDWAF_ERROR("{}", e.what());
     } catch (...) {
